rename logging and scraperwiki - module name clash on python 2.7 [#1]
* rename logging to hdx_logging and scraperwiki to hdx_scraperwiki
