+++ hdx/facades/hdx_scraperwiki.py
@@ -8,7 +8,7 @@ import scraperwiki
 
 from hdx.configuration import Configuration
 from hdx.facades import logging_kwargs
-from hdx.logging import setup_logging
+from hdx.hdx_logging import setup_logging
 
 logger = logging.getLogger(__name__)
 setup_logging(**logging_kwargs)
